Fix several installation issues, like LICENSE.txt missing.
This adds the obviously missing MANIFEST.in, but also fixes a couple otther
issues.

The perils of Python installation and distribution.  I need LICENSE.txt at
runtime, but I always must have that file at the root of the repo.  It turns
out, this is basically not supported.  So, I created a symlink to the
LICENSE.txt file so that package_data will pick it up and install it.